consolidate source files in top-level package zed
This commit consolidates the varied sources files in the top-level
zed package into fewer files, in part, to reduce the clutter in the
github repository listing.  There are a few more moves to make but
this is a start.
