--- array.go
@@ -1,18 +1,0 @@
-package zed
-
-type TypeArray struct {
-	id   int
-	Type Type
-}
-
-func NewTypeArray(id int, typ Type) *TypeArray {
-	return &TypeArray{id, typ}
-}
-
-func (t *TypeArray) ID() int {
-	return t.id
-}
-
-func (t *TypeArray) Kind() Kind {
-	return ArrayKind
-}
