--- null.go
@@ -1,11 +1,0 @@
-package zed
-
-type TypeOfNull struct{}
-
-func (t *TypeOfNull) ID() int {
-	return IDNull
-}
-
-func (t *TypeOfNull) Kind() Kind {
-	return PrimitiveKind
-}
